+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.1.0 (2012-09-28)
+
+[CHANGED] Renamed get_channel_stats to get_channel_info
+[CHANGED] get_channels now takes and $options parameter. get_channels( $options = array() )
+[REMOVED] get_presence_channels
+
 ## 2.0.1 (2012-09-18)
 
 [FIXED] Overwritten socket_id parameter in trigger: https://github.com/pusher/pusher-php-server/pull/3